Update) BookTable 최근 검색 항목, 유저 북마크 Attribute 대응 항목 Migration
